+++ NEWS.md
@@ -1,5 +1,4 @@
-
-# Packrat 0.7.0 (UNRELEASED)
+# Packrat 0.7.0
 
 - Fixed an issue where Packrat could inadvertently execute non-R code chunks
   when scanning R Markdown documents for dependencies.